test infrastructure

+++ 02_data_structures/dynamic_array.hpp
@@ -1,6 +1,8 @@
 #ifndef DYNAMIC_ARRAY
 #define DYNAMIC_ARRAY
 
+#include <cstddef>
+
 template <typename T>
 class DynamicArray
 {
@@ -8,11 +10,11 @@ class DynamicArray
     // добавляет элемент в конец массива
     virtual void   add(T item) = 0;
     // вставляет элемент по индексу со сдвигом хвоста массива вправо
-    virtual void   add(T item, int index) = 0;
+    virtual void   add(T item, size_t index) = 0;
     // извлекает элемент массива по индексу
-    virtual T      get(int index) const = 0;
+    virtual T      get(size_t index) const = 0;
     // удаляет элемент массива со сдвигом хвоста массива влево 
-    virtual T      remove(int index) = 0;
+    virtual T      remove(size_t index) = 0;
     // возвращает актуальный размер массива 
     virtual size_t size() const = 0;
 };
